+++ test/jest/helpers/getHookExecutionResult.js
@@ -1,11 +1,18 @@
 import React from 'react';
+import { isFunction } from 'lodash';
 
 import { renderWithIntl } from './renderWithIntl';
 
 export const getHookExecutionResult = (hook, hookArguments = []) => {
-  const result = {};
+  let result = {};
   const TestComponent = () => {
-    Object.assign(result, hook(...hookArguments));
+    const hookResult = hook(...hookArguments);
+
+    if (isFunction(hookResult)) {
+      result = hookResult;
+    } else {
+      Object.assign(result, hookResult);
+    }
 
     return null;
   };